Update install scripts to install docker distribution of OpenCTI
This also has the connectors.sh exit early, as this logic is now moved
into the installer.sh file. Have it set up all the docker images, by
pulling what's already built for arm64 off of dockerhub, and building
locally the connectors, many of which are not hosted on dockerhub for
the amd64 target. Also force a dist-upgrade of Ubuntu early on, and
start using the official docker-ce packages, rather than the older ones
from Ubuntu's main package repository.